Refactor playback module and UI components
- Updated imports in `player.rs` to use the new event module.
- Changed `Input` enum to `AudioInput` type alias for clarity.
- Refactored `run` function in `player.rs` to accept `SharedReader` directly.
- Modified `PlaylistQueue` in `queue.rs` to remove unnecessary methods and added new methods for song management.
- Introduced `scan.rs` for scanning local music files and extracting metadata.
- Updated `source.rs` to reflect changes in `player.rs` and improved error handling.
- Enhanced `PlaybackState` in `types.rs` with new methods for managing playback state.
- Expanded `state.rs` to include a new `ContentState` enum for better state management.
- Refactored UI components to use the new `config::Theme` instead of `theme::Theme`.
- Updated various UI files to improve consistency and maintainability.
