+++ src/event.rs
@@ -1,8 +1,77 @@
-pub use crate::types::{AppEvent, Event};
+use std::time::Duration;
 
+use crossterm::event::Event as CrosstermEvent;
 use color_eyre::eyre::OptionExt;
 use futures::{FutureExt, StreamExt};
+use ncm_api::{LoginInfo, SongInfo};
 use tokio::sync::mpsc;
+
+use crate::playback::types::LyricLine;
+use crate::state::{CommandAction, Page, SplashLogEntry};
+use crate::state::ContentState;
+
+#[derive(Clone, Debug)]
+pub enum Event {
+    Crossterm(CrosstermEvent),
+    App(AppEvent),
+}
+
+#[derive(Clone, Debug)]
+pub enum AppEvent {
+    Quit,
+    SplashTick {
+        progress: f64,
+        log: Option<SplashLogEntry>,
+    },
+    Login,
+    LoginSuccess(LoginInfo),
+    LoginError(String),
+    CaptchaSent,
+    QRCreated {
+        url: String,
+        key: String,
+    },
+    QRStatus(String),
+    NavSelect(String),
+    ContentLoaded(ContentState),
+    PlaylistSelect { id: u64, name: Option<String> },
+    BreadcrumbSet(String),
+    SongPlay(u64),
+    PlaybackStarted,
+    PlaybackProgress {
+        position: Duration,
+        total: Option<Duration>,
+    },
+    PlaybackFinished,
+    PlaybackError(String),
+    LyricsLoaded {
+        song_id: u64,
+        lyrics: Vec<LyricLine>,
+        translated_lyrics: Vec<LyricLine>,
+    },
+    HeartbeatSong(SongInfo),
+    HeartbeatFallback,
+    SearchSong(String),
+    LocalMusicLoaded(Vec<SongInfo>),
+    SetOffline,
+    Navigate(Page),
+    CommandPanel(CommandPanelAction),
+    SearchActivated,
+    SearchDeactivated,
+    ToggleBordered,
+    ExecuteCommand(CommandAction),
+    ContentRestore,
+    CellAction(usize, usize),
+}
+
+#[derive(Clone, Debug)]
+pub enum CommandPanelAction {
+    Open,
+    Close,
+    Previous,
+    Next,
+    Select,
+}
 
 pub struct EventHandler {
     sender: mpsc::UnboundedSender<Event>,
