+++ src/playback.rs
@@ -1,45 +1,42 @@
 mod controller;
+mod heartbeat;
 mod mode;
 mod player;
 mod queue;
+mod scan;
 mod source;
 mod storage;
 pub mod types;
 
-use std::fs;
-use std::io::BufReader;
 use std::path::PathBuf;
 use std::sync::Arc;
 use std::time::Duration;
 
 use ncm_api::{NcmClient, SongInfo};
-use rodio::Source;
 use tokio::sync::mpsc;
 
-use crate::types::{AppEvent, Event};
+use crate::event::{AppEvent, Event};
 
-use self::controller::PlayerController;
+use self::controller::PlaybackHandle;
 use self::mode::PlayStrategy;
 use self::queue::PlaylistQueue;
 use self::source::AudioSource;
 use self::storage::PlaylistStorage;
 use self::types::{LyricLine, PlayMode, PlaybackState};
 
-pub use self::types::{
-    LyricLine as PlaybackLyricLine, PlayMode as EnginePlayMode, PlaybackState as EngineState,
-    parse_lyric_lines,
-};
+pub use self::types::{PlayMode as EnginePlayMode, PlaybackState as EngineState};
 
 pub struct PlaybackEngine {
     pub state: PlaybackState,
-    pub queue: PlaylistQueue,
+    pub(super) queue: PlaylistQueue,
     strategy: Box<dyn PlayStrategy>,
     storage: PlaylistStorage,
-    pub source: AudioSource,
-    controller: PlayerController,
-    event_tx: mpsc::UnboundedSender<Event>,
-    api: Arc<NcmClient>,
+    source: AudioSource,
+    pub(super) controller: PlaybackHandle,
+    pub(super) event_tx: mpsc::UnboundedSender<Event>,
+    pub(super) api: Arc<NcmClient>,
     playlist_id: Option<u64>,
+    consecutive_errors: u32,
 }
 
 impl PlaybackEngine {
@@ -56,10 +53,11 @@ impl PlaybackEngine {
             strategy: Box::new(mode::Sequential),
             storage,
             source: AudioSource::new(api.clone(), cache_dir),
-            controller: PlayerController::new(event_tx.clone()),
+            controller: PlaybackHandle::new(event_tx.clone()),
             event_tx: event_tx.clone(),
             api,
             playlist_id: None,
+            consecutive_errors: 0,
         };
         this.restore_session();
         this
@@ -67,6 +65,24 @@ impl PlaybackEngine {
 
     pub fn current_song(&self) -> Option<SongInfo> {
         self.queue.current_song().cloned()
+    }
+
+    pub fn is_currently_playing(&self, song_id: u64) -> bool {
+        self.state.playing
+            && self
+                .state
+                .current_song
+                .as_ref()
+                .is_some_and(|s| s.id == song_id)
+    }
+
+    pub fn finish_and_snapshot(&mut self) -> Option<(u64, u64, f64)> {
+        let info = self.state.current_song.as_ref().map(|s| {
+            let progress = self.state.progress;
+            (s.id, s.duration, progress)
+        });
+        self.handle_finished();
+        info
     }
 
     pub fn song_at(&self, index: usize) -> Option<&SongInfo> {
@@ -83,6 +99,18 @@ impl PlaybackEngine {
 
     pub fn queue_history(&self) -> &[SongInfo] {
         &self.queue.history
+    }
+
+    pub fn queue_current_index(&self) -> Option<usize> {
+        self.queue.current_index
+    }
+
+    pub fn set_queue_songs(&mut self, songs: Vec<SongInfo>) {
+        self.queue.songs = songs;
+    }
+
+    pub fn cache(&self) -> &crate::cache::CacheManager {
+        &self.source.cache
     }
 
     pub fn play_songs(&mut self, songs: Vec<SongInfo>, index: usize) {
@@ -113,12 +141,7 @@ impl PlaybackEngine {
             return;
         }
         self.controller.stop();
-        if let Some(old_idx) = self.queue.current_index
-            && let Some(song) = self.queue.songs.get(old_idx)
-        {
-            self.queue.history.push(song.clone());
-        }
-        self.queue.current_index = Some(index);
+        self.queue.advance_to(index);
         self.strategy =
             mode::create_strategy(&self.state.mode, self.queue.len(), self.queue.current_index);
         self.start_current_song(None);
@@ -130,22 +153,13 @@ impl PlaybackEngine {
         }
 
         if matches!(self.state.mode, PlayMode::Heartbeat { .. }) {
-            // FIXME(Heartbeat): `next_heartbeat` 仅 spawn 异步请求并立即返回，不会停止当前歌曲，
-            // 也不会推进 current_index。当前歌曲播完后 on_playback_finished → next() 会再次
-            // next_heartbeat()，导致多个 in-flight 请求交错、可能拿到过期/重复的下一首。
             self.next_heartbeat();
             return;
         }
 
-        let next_idx = self
-            .strategy
-            .next(self.queue.current_index, &self.queue.songs);
-        match next_idx {
+        match self.queue.next_index(&mut *self.strategy) {
             Some(idx) => {
-                if Some(idx) != self.queue.current_index {
-                    self.queue.push_to_history();
-                }
-                self.queue.current_index = Some(idx);
+                self.queue.advance_to(idx);
                 self.start_current_song(None);
             }
             None => {
@@ -167,10 +181,7 @@ impl PlaybackEngine {
             return;
         }
 
-        let prev_idx = self
-            .strategy
-            .prev(self.queue.current_index, &self.queue.songs);
-        if let Some(idx) = prev_idx {
+        if let Some(idx) = self.queue.prev_index(&mut *self.strategy) {
             self.queue.current_index = Some(idx);
             self.start_current_song(None);
         }
@@ -264,48 +275,36 @@ impl PlaybackEngine {
     }
 
     pub fn handle_finished(&mut self) {
-        if !self.state.playing {
-            return;
+        let should_advance = self.state.on_finished();
+        if should_advance {
+            self.next();
         }
-        self.next();
+        if !self.state.playing {
+            self.state.clear_after_stopped();
+        }
     }
 
     pub fn on_playback_started(&mut self) {
-        self.state.error = None;
-        self.state.paused = false;
-        self.state.playing = true;
-        self.state.lyrics = None;
-        self.state.translated_lyrics = None;
+        self.state.on_started();
+        self.consecutive_errors = 0;
     }
 
     pub fn on_playback_progress(&mut self, position: Duration, total: Option<Duration>) {
-        self.state.seeking = false;
-        let total_secs = match total {
-            Some(t) => t.as_secs_f64(),
-            None => self
-                .queue
-                .current_song()
-                .map(|s| s.duration as f64 / 1000.0)
-                .unwrap_or(0.0),
-        };
-        if total_secs > 0.0 {
-            self.state.progress = (position.as_secs_f64() / total_secs).clamp(0.0, 1.0);
-        }
+        self.state.on_progress(position, total);
     }
 
     pub fn on_playback_finished(&mut self) {
-        self.state.progress = 0.0;
         self.handle_finished();
-        if !self.state.playing {
-            self.state.current_song = None;
-            self.state.error = None;
-            self.state.paused = false;
-        }
     }
 
     pub fn on_playback_error(&mut self, err: String) {
-        log::error!("Playback error: {}", err);
-        self.state.error = Some(err);
+        self.state.on_error(err);
+        self.consecutive_errors += 1;
+        if self.consecutive_errors >= 3 {
+            self.stop();
+        } else {
+            self.next();
+        }
     }
 
     pub fn on_lyrics_loaded(
@@ -314,12 +313,8 @@ impl PlaybackEngine {
         lyrics: Vec<LyricLine>,
         translated_lyrics: Vec<LyricLine>,
     ) {
-        if let Some(song) = &self.state.current_song
-            && song.id == song_id
-        {
-            self.state.lyrics = Some(lyrics);
-            self.state.translated_lyrics = Some(translated_lyrics);
-        }
+        self.state
+            .on_lyrics_loaded(song_id, lyrics, translated_lyrics);
     }
 
     pub fn save_session(&self) {
@@ -377,7 +372,7 @@ impl PlaybackEngine {
         }
     }
 
-    fn start_current_song(&mut self, seek_time: Option<Duration>) {
+    pub(super) fn start_current_song(&mut self, seek_time: Option<Duration>) {
         let song = match self.queue.current_song() {
             Some(s) => Arc::new(s.clone()),
             None => return,
@@ -425,123 +420,6 @@ impl PlaybackEngine {
         });
     }
 
-    // FIXME(Heartbeat): 已知问题 ——
-    // 1. 当前歌曲播完前不会 stop，会重复触发 next_heartbeat 产生并发请求；
-    // 2. playmode_intelligence_list 返回列表只取第一首(next())，其余丢弃，
-    //    没有真正维护"心动队列"，每次切歌都再发一次网络请求；
-    // 3. HeartbeatSong 异步回传与 on_playback_finished 之间存在竞态，
-    //    快速切歌时可能播放到过时推荐。
-    fn next_heartbeat(&mut self) {
-        let song = match self.queue.current_song() {
-            Some(s) => s.clone(),
-            None => return,
-        };
-        let playlist_id = match self.state.mode {
-            PlayMode::Heartbeat { playlist_id } => playlist_id,
-            _ => return,
-        };
-
-        log::info!("Heartbeat: song_id={} playlist_id={}", song.id, playlist_id);
-
-        let api = self.api.clone();
-        let event_tx = self.event_tx.clone();
-
-        tokio::spawn(async move {
-            match api.playmode_intelligence_list(song.id, playlist_id).await {
-                Ok(songs) => {
-                    // FIXME(Heartbeat): 只取推荐列表第一首，其余丢弃，没有维护"心动队列"。
-                    // 建议缓存整段推荐队列，本地推进，仅在队列耗尽时再请求。
-                    if let Some(next_song) = songs.into_iter().next() {
-                        if event_tx
-                            .send(Event::App(AppEvent::HeartbeatSong(next_song)))
-                            .is_err()
-                        {
-                            log::error!("Failed to send HeartbeatSong: receiver dropped");
-                        }
-                    } else if event_tx
-                        .send(Event::App(AppEvent::HeartbeatFallback))
-                        .is_err()
-                    {
-                        log::error!("Failed to send HeartbeatFallback: receiver dropped");
-                    }
-                }
-                Err(e) => {
-                    log::warn!("Heartbeat failed: {e}, falling back to queue");
-                    if event_tx
-                        .send(Event::App(AppEvent::HeartbeatFallback))
-                        .is_err()
-                    {
-                        log::error!("Failed to send HeartbeatFallback: receiver dropped");
-                    }
-                }
-            }
-        });
-
-        // NOTE(Heartbeat): 故意保持 state.playing=true，避免 on_playback_finished 清掉
-        // current_song。但这样会与上面 FIXME 的竞态叠加：歌曲已结束、progress 已归零、
-        // 下一首尚未送达期间，用户操作或再次触发 next 会产生未定义行为。
-    }
-
-    // FIXME(Heartbeat): 从 AppEvent(HeartbeatSong) 异步回调进入，先 stop 再 push 新歌。
-    // 多个并发的 next_heartbeat 请求可能依次送达，导致一首歌被"下一首"覆盖多次，
-    // 或播放顺序与用户预期不符。
-    pub fn play_heartbeat_song(&mut self, song: SongInfo) {
-        self.controller.stop();
-        self.queue.push_to_history();
-        self.queue.songs.push(song);
-        self.queue.current_index = Some(self.queue.len() - 1);
-        self.start_current_song(None);
-    }
-
-    pub fn on_heartbeat_fallback(&mut self) {
-        log::info!("Heartbeat fallback: switching to sequential");
-        self.set_mode(PlayMode::Sequential);
-        self.next();
-    }
 }
 
-pub fn scan_local_music(dir: &std::path::Path) -> Vec<SongInfo> {
-    let Ok(entries) = fs::read_dir(dir) else {
-        return vec![];
-    };
-    let extensions = ["mp3", "flac", "wav", "ogg", "aac", "m4a", "wma"];
-    let mut songs = Vec::new();
-    for entry in entries.filter_map(|e| e.ok()) {
-        let path = entry.path();
-        if !path.is_file() {
-            continue;
-        }
-        let ext = path
-            .extension()
-            .and_then(|e| e.to_str())
-            .map(|e| e.to_lowercase())
-            .unwrap_or_default();
-        if !extensions.contains(&ext.as_str()) {
-            continue;
-        }
-        let name = path
-            .file_stem()
-            .and_then(|n| n.to_str())
-            .unwrap_or("unknown")
-            .to_string();
-        let duration = fs::File::open(&path)
-            .ok()
-            .and_then(|f| rodio::Decoder::new(BufReader::new(f)).ok())
-            .and_then(|d| d.total_duration())
-            .map(|d| d.as_millis() as u64)
-            .unwrap_or(0);
-        songs.push(SongInfo {
-            id: 0,
-            name,
-            singer: "本地".into(),
-            artist_id: 0,
-            album: path.to_string_lossy().to_string(),
-            album_id: 0,
-            pic_url: String::new(),
-            duration,
-            copyright: ncm_api::SongCopyright::Free,
-        });
-    }
-    songs.sort_by(|a, b| a.name.cmp(&b.name));
-    songs
-}
+pub use self::scan::scan_local_music;
