+++ src/ui.rs
@@ -29,7 +29,7 @@ use self::block::CornerBlock;
 use crate::{
     layout,
     state::{App, Page},
-    theme::Theme,
+    config::Theme,
 };
 
 pub fn calc_scroll_offset(selected: usize, visible_height: usize, total: usize) -> usize {
